Update libraries
Ionic 1.1.1, now has built in support for iOS9
ui-leaflet, a fork of angular-leaflet-directive
Also removed some unused libraries and such.
From now on, bower.json will contain the full version of libraries,
e.g. "1.5.0" instead of "~1.5.0" or "1.5.x".
This is so we will not update libraries by mistake, and break something.