fix(ios): update ionic-webview to v4
resolves issues with never ending splash screens

BREAKING CHANGE: Removes the localStorage migration plugin. Everyone is
going to have to log in again. It sucks, but what can you do, the plugin
is too much work to rewrite for this.

BREAKING CHANGE: drop support for iOS < 11

BREAKING CHANGE: The app is now served from ifiske://localhost in order
to migitage some security issues with the previous model. This will
require changes to the CORS settings in the API.